Separating out for composite projections
